docs: reorganize combinatorics documentation and refresh README

Add dedicated vignettes for integer partitions and integer compositions
and refactor the constraints vignette to focus on constraint-driven
enumeration.

Improve README messaging and structure to better highlight key package
features and unique algorithms (e.g. multiset partitions, comboGroups,
comboGrid, and distinct compositions).

Switch vignette output to prettydoc::html_pretty and update vignette
metadata. Regenerate documentation outputs.

Minor repository housekeeping:

* move CRAN release checklist to dev/
* ignore dev/ in .Rbuildignore

+++ src/CompositionsDistinct.cpp
@@ -407,7 +407,7 @@ int CompsDistinct(int* mat, std::vector<int> &z, std::size_t width,
                 }
             }
 
-            if (nextStep >= static_cast<int>(nRows)) return 1;
+            if (nextStep >= nRows) return 1;
             strt = nextStep;
 
             std::iota(z.begin(), z.end(), 1);
@@ -477,7 +477,7 @@ int CompsGenDistinct(
                 }
             }
 
-            if (nextStep >= static_cast<int>(nRows)) return 1;
+            if (nextStep >= nRows) return 1;
             strt = nextStep;
 
             if (idx_max == tar) {
@@ -550,7 +550,7 @@ int CompsDistinct(
                 }
             }
 
-            if (nextStep >= static_cast<int>(nRows)) return 1;
+            if (nextStep >= nRows) return 1;
             strt = nextStep;
 
             std::iota(z.begin(), z.end(), 1);
@@ -619,7 +619,7 @@ int CompsGenDistinct(
                 }
             }
 
-            if (nextStep >= static_cast<int>(nRows)) return 1;
+            if (nextStep >= nRows) return 1;
             strt = nextStep;
 
             if (idx_max == tar) {
